Factored out code that can be reused in examples
