+++ samples/util/util.h
@@ -1,0 +1,68 @@
+/**
+* @version		GrPPI v0.2
+* @copyright		Copyright (C) 2017 Universidad Carlos III de Madrid. All rights reserved.
+* @license		GNU/GPL, see LICENSE.txt
+* This program is free software: you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation, either version 3 of the License, or
+* (at your option) any later version.
+*
+* This program is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU General Public License for more details.
+*
+* You have received a copy of the GNU General Public License in LICENSE.txt
+* also available in <http://www.gnu.org/licenses/gpl.html>.
+*
+* See COPYRIGHT.txt for copyright notices and details.
+*/
+
+#ifndef GRPPI_SAMPLES_UTIL_H
+#define GRPPI_SAMPLES_UTIL_H
+
+grppi::polymorphic_execution execution_mode(const std::string & opt) {
+  using namespace grppi;
+  if ("seq" == opt) 
+    return make_polymorphic_execution<sequential_execution>();
+  if ("thr" == opt) 
+    return make_polymorphic_execution<parallel_execution_thr>();
+  if ("omp" == opt) 
+    return make_polymorphic_execution<parallel_execution_omp>();
+  if ("tbb" == opt) 
+    return make_polymorphic_execution<parallel_execution_tbb>();
+  return polymorphic_execution{};
+}
+
+template <typename F, typename ...Args>
+bool run_test(const string & mode, F && f, Args && ... args) {
+  auto e = execution_mode(mode);
+  if (e.has_execution()) {
+    f(e, std::forward<Args>(args)...);
+    return true;
+  }
+  return false;
+}
+
+void print_available_modes(std::ostream & os) {
+  using namespace std;
+  using namespace grppi;
+
+  if (is_supported<sequential_execution>()) {
+    os << "    seq -> Sequential execution" << endl;
+  }
+
+  if (is_supported<parallel_execution_thr>()) {
+    os << "    thr -> ISO Threads backend" << endl;
+  }
+
+  if (is_supported<parallel_execution_tbb>()) {
+    os << "    tbb -> Intel TBB backend" << endl;
+  }
+
+  if (is_supported<parallel_execution_omp>()) {
+    os << "    omp -> OpenMP backend" << endl;
+  }
+}
+
+#endif
